again

+++ 0x0A-argc_argv/4-add.c
@@ -11,7 +11,7 @@
 int main(int argc, char *argv[])
 {
 	int i;
-	unsigned int sum;
+	long int sum;
 
 	for (i = 1; i < argc; i++)
 	{
@@ -25,6 +25,6 @@ int main(int argc, char *argv[])
 			return (1);
 		}
 	}
-	printf("%d\n", sum);
+	printf("%lu\n", sum);
 	return (sum);
 }
